pyproject.toml
[project]
name = "lenny"
version = "0.1.0"
description = "Lenny project created with PDM"
authors = [
	{ name = "Jane Doe", email = "jane.doe@example.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
keywords = ["example", "pdm", "lenny"]
classifiers = [
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",

]

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "esipy>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.11.0",
    "mcp>=1.0.0",
    "sse-starlette>=1.6.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "opentelemetry-distro>=0.40b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "opentelemetry-instrumentation-celery>=0.40b0",
    "fastapi-pagination>=0.12.0",
]

[project.urls]
Home = "https://example.com/lenny"
Documentation = "https://example.com/lenny/docs"
Source = "https://example.com/lenny/source"
Tracker = "https://example.com/lenny/issues"

[project.scripts]
lenny = "lenny.main:main"

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["pdm-backend>=0.10"]
build-backend = "pdm.backend"
